Added payment and activity functionality
When an activity is chosen, activities at the same time can no longer be
selected.  The price of the chosen activity is added to the bottom of
the activities list.  Details of a payment method will not be shown if
the payment method is not selected from the payment dropdown.